move misc implementation out of `jit/__init__.py`
[ghstack-poisoned]
