+++ test/quantization/test_quantize.py
@@ -1642,7 +1642,7 @@ class TestDeprecatedJitQuantized(JitTestCase):
             def weight(self, w):
                 self._packed_weight = torch.ops.quantized.linear_prepack(w)
 
-        with torch.jit._disable_emit_hooks():
+        with torch._jit_internal._disable_emit_hooks():
             x = torch.jit.script(Linear(10, 10))
             torch._C._jit_pass_erase_shape_information(x.graph)
